requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
sqlalchemy>=2.0.38
pydantic-settings>=2.8.0
pytest>=8.3.5
httpx>=0.28.1
